python/requirements.txt
Flask
mysql-connector-python
pyjwt
